# ragbag-api/openapi.yaml
openapi: 3.1.0
info:
  title: Authentication Api
  version: 1.0.0

paths:
  # Authentication
  /auth/register:
    post:
      summary: Register a new user
      operationId: register
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '409':
          $ref: '#/components/responses/Conflict'

  /auth/login:
    post:
      summary: Login
      operationId: login
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: OK
          headers:
            Set-Cookie:
              description: >
                HttpOnly refresh token cookie.
                Path=/api/auth/refresh; SameSite=Lax; Secure (in prod).
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'

  /auth/refresh:
    post:
      summary: Refresh access token
      operationId: refresh
      description: >
        Uses the HttpOnly refresh_token cookie to issue a new access token.
      tags:
        - Authentication
      parameters:
        - in: cookie
          name: refresh_token
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /auth/logout/all:
    get:
      summary: Logout from all devices
      operationId: logoutAll
      tags:
        - Authentication
      security:
        - bearerAuth: [ ]
      responses:
        '203':
          $ref: '#/components/responses/NoContent'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # Users
  /v1/users/me:
    get:
      summary: Get current user
      operationId: me
      tags:
        - User
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # Spaces
  /v1/spaces:
    get:
      summary: Get all spaces
      operationId: getSpaces
      tags:
        - Space
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpaceDtoList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  /v1/spaces/{spaceId}:
    get:
      summary: Get a space
      operationId: getSpace
      tags:
        - Space
      security:
        - bearerAuth: [ ]
      parameters:
        - in: path
          name: spaceId
          required: true
          schema:
            $ref: '#/components/schemas/ID'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpaceDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # Collections
  /v1/spaces/{spaceId}/collections:
    get:
      summary: Get all collections of a space
      operationId: getCollections
      tags:
        - Collection
      security:
        - bearerAuth: [ ]
      parameters:
        - in: path
          name: spaceId
          required: true
          schema:
            $ref: '#/components/schemas/ID'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionDtoList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      summary: Create a collection
      operationId: createCollection
      tags:
        - Collection
      security:
        - bearerAuth: [ ]
      parameters:
        - in: path
          name: spaceId
          required: true
          schema:
            $ref: '#/components/schemas/ID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCollectionDto'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /v1/collections/{collectionId}:
    get:
      summary: Get a collection
      operationId: getCollection
      tags:
        - Collection
      security:
        - bearerAuth: [ ]
      parameters:
        - in: path
          name: collectionId
          required: true
          schema:
            $ref: '#/components/schemas/ID'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update a collection
      operationId: updateCollection
      tags:
        - Collection
      security:
        - bearerAuth: [ ]
      parameters:
        - in: path
          name: collectionId
          required: true
          schema:
            $ref: '#/components/schemas/ID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCollectionDto'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete a collection
      operationId: deleteCollection
      tags:
        - Collection
      security:
        - bearerAuth: [ ]
      parameters:
        - in: path
          name: collectionId
          required: true
          schema:
            $ref: '#/components/schemas/ID'
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # Bookmarks
  /v1/collections/{collectionId}/bookmarks:
    get:
      summary: Get all bookmarks of a collection
      operationId: getBookmarks
      tags:
        - Bookmark
      security:
        - bearerAuth: [ ]
      parameters:
        - in: path
          name: collectionId
          required: true
          schema:
            $ref: '#/components/schemas/ID'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookmarkDtoList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      summary: Create a bookmark
      operationId: createBookmark
      tags:
        - Bookmark
      security:
        - bearerAuth: [ ]
      parameters:
        - in: path
          name: collectionId
          required: true
          schema:
            $ref: '#/components/schemas/ID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookmarkDto'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookmarkDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /v1/bookmarks/{bookmarkId}:
    get:
      summary: Get a bookmark
      operationId: getBookmark
      tags:
        - Bookmark
      security:
        - bearerAuth: [ ]
      parameters:
        - in: path
          name: bookmarkId
          required: true
          schema:
            $ref: '#/components/schemas/ID'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookmarkDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update a bookmark
      operationId: updateBookmark
      tags:
        - Bookmark
      security:
        - bearerAuth: [ ]
      parameters:
        - in: path
          name: bookmarkId
          required: true
          schema:
            $ref: '#/components/schemas/ID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBookmarkDto'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookmarkDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete a bookmark
      operationId: deleteBookmark
      tags:
        - Bookmark
      security:
        - bearerAuth: [ ]
      parameters:
        - in: path
          name: bookmarkId
          required: true
          schema:
            $ref: '#/components/schemas/ID'
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # Invites
  /api/invites:
    post:
      summary: Create an invite link
      tags: 
        - Invite
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInviteRequest'
      responses:
        '201':
          description: Invite created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InviteDto'

    get:
      summary: List invites created by current user
      tags: 
        - Invite
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: List of invites
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InviteDto'

  /api/invites/{token}:
    post:
      summary: Accept an invite
      tags:
        - Invite
      security:
        - bearerAuth: [ ]
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Invite accepted
        '400':
          description: Invite invalid or expired
        '409':
          description: Invite already used

    delete:
      summary: Revoke an invite
      tags:
        - Invite
      security:
        - bearerAuth: [ ]
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Invite revoked

components:
  schemas:
    ID:
      type: string
      description: UUID
      example: "fa2b6a4a-5403-498c-9be8-c31d7d01a724"
      minLength: 36
      maxLength: 36

    JWT:
      type: string
      description: JSON Web Token
      example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdW..."

    Error:
      type: object
      properties:
        message:
          type: string

    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password

    LoginResponse:
      type: object
      properties:
        access_token:
          $ref: '#/components/schemas/JWT'
      required:
        - access_token

    RegisterRequest:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
      required:
        - username
        - email
        - password

    UserDto:
      properties:
        id:
          $ref: '#/components/schemas/ID'
        username:
          type: string
        email:
          type: string
      required:
        - id
        - username
        - email

    SpaceDto:
      properties:
        id:
          $ref: '#/components/schemas/ID'
        name:
          type: string
        description:
          type: string
      required:
        - id
        - name

    SpaceDtoList:
      type: array
      items:
        $ref: '#/components/schemas/SpaceDto'

    CollectionDto:
      properties:
        id:
          $ref: '#/components/schemas/ID'
        name:
          type: string
        description:
          type: string
        parentId:
          $ref: '#/components/schemas/ID'
        children:
          type: array
          items:
            $ref: '#/components/schemas/CollectionDto'
      required:
        - id
        - name

    CollectionVisibilityDto:
      type: string
      enum:
        - PRIVATE
        - INTERNAL
        - PUBLIC

    CollectionDtoList:
      type: array
      items:
        $ref: '#/components/schemas/CollectionDto'

    CreateCollectionDto:
      properties:
        name:
          type: string
      required:
        - name

    UpdateCollectionDto:
      properties:
        name:
          type: string
      required:
        - name

    BookmarkDto:
      properties:
        id:
          $ref: '#/components/schemas/ID'
        collectionId:
          $ref: '#/components/schemas/ID'
        url:
          type: string
        title:
          type: string
        description:
          type: string
      required:
        - id
        - url

    BookmarkDtoList:
      type: array
      items:
        $ref: '#/components/schemas/BookmarkDto'

    CreateBookmarkDto:
      properties:
        url:
          type: string
        title:
          type: string
        description:
          type: string
      required:
        - url

    UpdateBookmarkDto:
      properties:
        url:
          type: string
        title:
          type: string
        description:
          type: string
      required:
        - url

    InviteDto:
      type: object
      properties:
        token:
          type: string
        targetType:
          type: string
          enum: [ SPACE, COLLECTION ]
        targetId:
          type: string
        role:
          type: string
          enum: [ VIEWER, CONTRIBUTOR ]
        expiresAt:
          type: string
          format: date-time
          nullable: true
        maxUses:
          type: integer
          nullable: true
        usedCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        invitedBy:
          type: string

    CreateInviteRequest:
      type: object
      required:
        - targetType
        - targetId
      properties:
        targetType:
          type: string
          enum: [ SPACE, COLLECTION ]
        targetId:
          type: string
        role:
          type: string
          enum: [ VIEWER, CONTRIBUTOR ]
          default: VIEWER
        expiresAt:
          type: string
          format: date-time
          nullable: true
        maxUses:
          type: integer
          nullable: true

  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NoContent:
      description: No Content
    Unauthorized:
      description: Unauthorized
    Forbidden:
      description: Forbidden
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Conflict

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

security:
  - bearerAuth: []