# ragbag-frontend/src/lib/api/gen/models/CreateInviteRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Authentication Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreateInviteRequest
 */
export interface CreateInviteRequest {
    /**
     * 
     * @type {string}
     * @memberof CreateInviteRequest
     */
    targetType: CreateInviteRequestTargetTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof CreateInviteRequest
     */
    targetId: string;
    /**
     * 
     * @type {string}
     * @memberof CreateInviteRequest
     */
    role?: CreateInviteRequestRoleEnum;
    /**
     * 
     * @type {Date}
     * @memberof CreateInviteRequest
     */
    expiresAt?: Date;
    /**
     * 
     * @type {number}
     * @memberof CreateInviteRequest
     */
    maxUses?: number;
}


/**
 * @export
 */
export const CreateInviteRequestTargetTypeEnum = {
    Space: 'SPACE',
    Collection: 'COLLECTION'
} as const;
export type CreateInviteRequestTargetTypeEnum = typeof CreateInviteRequestTargetTypeEnum[keyof typeof CreateInviteRequestTargetTypeEnum];

/**
 * @export
 */
export const CreateInviteRequestRoleEnum = {
    Viewer: 'VIEWER',
    Contributor: 'CONTRIBUTOR'
} as const;
export type CreateInviteRequestRoleEnum = typeof CreateInviteRequestRoleEnum[keyof typeof CreateInviteRequestRoleEnum];


/**
 * Check if a given object implements the CreateInviteRequest interface.
 */
export function instanceOfCreateInviteRequest(value: object): value is CreateInviteRequest {
    if (!('targetType' in value) || value['targetType'] === undefined) return false;
    if (!('targetId' in value) || value['targetId'] === undefined) return false;
    return true;
}

export function CreateInviteRequestFromJSON(json: any): CreateInviteRequest {
    return CreateInviteRequestFromJSONTyped(json, false);
}

export function CreateInviteRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateInviteRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'targetType': json['targetType'],
        'targetId': json['targetId'],
        'role': json['role'] == null ? undefined : json['role'],
        'expiresAt': json['expiresAt'] == null ? undefined : (new Date(json['expiresAt'])),
        'maxUses': json['maxUses'] == null ? undefined : json['maxUses'],
    };
}

export function CreateInviteRequestToJSON(json: any): CreateInviteRequest {
    return CreateInviteRequestToJSONTyped(json, false);
}

export function CreateInviteRequestToJSONTyped(value?: CreateInviteRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'targetType': value['targetType'],
        'targetId': value['targetId'],
        'role': value['role'],
        'expiresAt': value['expiresAt'] == null ? undefined : ((value['expiresAt']).toISOString()),
        'maxUses': value['maxUses'],
    };
}