# ragbag-frontend/src/lib/api/gen/models/InviteDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Authentication Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface InviteDto
 */
export interface InviteDto {
    /**
     * 
     * @type {string}
     * @memberof InviteDto
     */
    token?: string;
    /**
     * 
     * @type {string}
     * @memberof InviteDto
     */
    targetType?: InviteDtoTargetTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof InviteDto
     */
    targetId?: string;
    /**
     * 
     * @type {string}
     * @memberof InviteDto
     */
    role?: InviteDtoRoleEnum;
    /**
     * 
     * @type {Date}
     * @memberof InviteDto
     */
    expiresAt?: Date;
    /**
     * 
     * @type {number}
     * @memberof InviteDto
     */
    maxUses?: number;
    /**
     * 
     * @type {number}
     * @memberof InviteDto
     */
    usedCount?: number;
    /**
     * 
     * @type {Date}
     * @memberof InviteDto
     */
    createdAt?: Date;
    /**
     * 
     * @type {string}
     * @memberof InviteDto
     */
    invitedBy?: string;
}


/**
 * @export
 */
export const InviteDtoTargetTypeEnum = {
    Space: 'SPACE',
    Collection: 'COLLECTION'
} as const;
export type InviteDtoTargetTypeEnum = typeof InviteDtoTargetTypeEnum[keyof typeof InviteDtoTargetTypeEnum];

/**
 * @export
 */
export const InviteDtoRoleEnum = {
    Viewer: 'VIEWER',
    Contributor: 'CONTRIBUTOR'
} as const;
export type InviteDtoRoleEnum = typeof InviteDtoRoleEnum[keyof typeof InviteDtoRoleEnum];


/**
 * Check if a given object implements the InviteDto interface.
 */
export function instanceOfInviteDto(value: object): value is InviteDto {
    return true;
}

export function InviteDtoFromJSON(json: any): InviteDto {
    return InviteDtoFromJSONTyped(json, false);
}

export function InviteDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): InviteDto {
    if (json == null) {
        return json;
    }
    return {
        
        'token': json['token'] == null ? undefined : json['token'],
        'targetType': json['targetType'] == null ? undefined : json['targetType'],
        'targetId': json['targetId'] == null ? undefined : json['targetId'],
        'role': json['role'] == null ? undefined : json['role'],
        'expiresAt': json['expiresAt'] == null ? undefined : (new Date(json['expiresAt'])),
        'maxUses': json['maxUses'] == null ? undefined : json['maxUses'],
        'usedCount': json['usedCount'] == null ? undefined : json['usedCount'],
        'createdAt': json['createdAt'] == null ? undefined : (new Date(json['createdAt'])),
        'invitedBy': json['invitedBy'] == null ? undefined : json['invitedBy'],
    };
}

export function InviteDtoToJSON(json: any): InviteDto {
    return InviteDtoToJSONTyped(json, false);
}

export function InviteDtoToJSONTyped(value?: InviteDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'token': value['token'],
        'targetType': value['targetType'],
        'targetId': value['targetId'],
        'role': value['role'],
        'expiresAt': value['expiresAt'] == null ? undefined : ((value['expiresAt']).toISOString()),
        'maxUses': value['maxUses'],
        'usedCount': value['usedCount'],
        'createdAt': value['createdAt'] == null ? undefined : ((value['createdAt']).toISOString()),
        'invitedBy': value['invitedBy'],
    };
}